+++ src/main/java/com/cobber/fta/StringFacts.java
@@ -1,7 +1,11 @@
 package com.cobber.fta;
 
+import java.util.SortedSet;
+
 public class StringFacts {
 	String minValue;
 	String maxValue;
 	String sum;
+	SortedSet<String> topK;
+	SortedSet<String> bottomK;
 }